% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{char}
\alias{char}
\title{A subset of data from kaggle web site: char
Report ...}
\format{
## `char`
a data frames containing: 62 rows and 11 columns

 @source <https://www.kaggle.com/datasets/bcruise/breaking-bad-episode-data>
}
\usage{
char
}
\description{
A subset of data from kaggle web site: char
Report ...
}
\keyword{datasets}
