% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{death}
\alias{death}
\title{A subset of data from kaggle web site: death
Report ...}
\format{
## `death`
a data frames containing: 65 rows and 8columns

 @source <https://www.kaggle.com/datasets/bcruise/breaking-bad-episode-data>
}
\usage{
death
}
\description{
A subset of data from kaggle web site: death
Report ...
}
\keyword{datasets}
