% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epi1}
\alias{epi1}
\title{Breaking Bad serie datas
A subset of data from kaggle web site:
Report ...
1st one epi1}
\format{
## `epi1`
a data frames containing 62 rows and 8 columns

\describe{

  ...
}
}
\source{
<https://www.kaggle.com/datasets/varpit94/breaking-bad-tv-show-all-seasons-episodes-data>
 <https://www.kaggle.com/datasets/parthhhpanchal/breaking-bad-datasets>
}
\usage{
epi1
}
\description{
Breaking Bad serie datas
A subset of data from kaggle web site:
Report ...
1st one epi1
}
\keyword{datasets}
