% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epi2}
\alias{epi2}
\title{A subset of data from kaggle web site: epi2
Report ...}
\format{
## `epi2`
a data frames containing: 102 rows and 7 columns

 @source <https://www.kaggle.com/datasets/bcruise/breaking-bad-episode-data>
}
\usage{
epi2
}
\description{
A subset of data from kaggle web site: epi2
Report ...
}
\keyword{datasets}
