% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{fonctions_metiers_propre_a_mon_api Title}
\usage{
say_hello(personne)
}
\arguments{
\item{personne}{la personne a qui on veut dire bjr}
}
\value{
character de type "Hey coucou {personne}"
}
\description{
Title: say hello
}
\examples{
say_hello(personne="Sarah")
say_hello(personne="Tchu")
}
