% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vote}
\alias{vote}
\title{A subset of data from kaggle web site: vote
Report ...}
\format{
## `vote`
a data frames containing: 62 rows and 7 columns

 @source <https://www.kaggle.com/datasets/bcruise/breaking-bad-episode-data>
}
\usage{
vote
}
\description{
A subset of data from kaggle web site: vote
Report ...
}
\keyword{datasets}
